feat(cal-ui-p4): move planning search components into package (generic)
Generic, domain-agnostic PlanningSearchAutocomplete + PlanningSearchTags now
live in @microboxlabs/miot-calendar-ui: the autocomplete groups matches by a
host-supplied searchable-field config (accessor + label + icon) and reads
chrome strings via the CalendarHost i18n seam; the tag list resolves match-type
labels via a host labelFor callback. Freight shims at the old app paths keep the
historical props/import paths and wire the freight field config + i18n.
